+++ frontend/apps/mobile/components/AvatarPicker.tsx
@@ -4,40 +4,20 @@ import { SvgXml } from "react-native-svg";
 import { ThemedText } from "@/components/themed-text";
 import { Colors } from "@/constants/theme";
 import { useColorScheme } from "@/hooks/use-color-scheme";
+import { useTranslation } from "react-i18next";
 import {
   UNIQUE_AVATAR_FACES,
   getSvgWithColor,
   getAvatarSvg,
 } from "@/constants/avatarFaces";
+import {
+  AVATAR_COLORS,
+  DARK_AVATAR_COLORS,
+  DEFAULT_AVATAR_COLOR,
+} from "@/constants/avatarColors";
 
 export { UNIQUE_AVATAR_FACES, getSvgWithColor, getAvatarSvg };
-
-export const AVATAR_COLORS = [
-  "#9CAF7D",
-  "#F5D878",
-  "#F5A85A",
-  "#B5A8E0",
-  "#F0A0BC",
-  "#E07070",
-  "#6B9FD4",
-  "#A8D4E8",
-  "#B0B0B0",
-];
-
-/** Darker shade of each AVATAR_COLOR for use as initials text */
-export const DARK_AVATAR_COLORS: Record<string, string> = {
-  "#9CAF7D": "#5E694B",
-  "#F5D878": "#938248",
-  "#F5A85A": "#936536",
-  "#B5A8E0": "#6D6586",
-  "#F0A0BC": "#906071",
-  "#E07070": "#864343",
-  "#6B9FD4": "#405F7F",
-  "#A8D4E8": "#657F8B",
-  "#B0B0B0": "#6A6A6A",
-};
-
-export const DEFAULT_AVATAR_COLOR = AVATAR_COLORS[3];
+export { AVATAR_COLORS, DARK_AVATAR_COLORS, DEFAULT_AVATAR_COLOR };
 
 const CIRCLE_SIZE = 62;
 const RING_WIDTH = 2.5;
@@ -98,6 +78,12 @@ export function AvatarPicker({
   const colorScheme = useColorScheme();
   const theme = Colors[colorScheme ?? "light"];
   const isDark = colorScheme === "dark";
+  const { t: translate } = useTranslation();
+
+  const TAB_LABELS: Record<Tab, string> = {
+    Colors: translate("childProfile.avatarTabColors"),
+    Avatar: translate("childProfile.avatarTabAvatar"),
+  };
 
   const selectedBg = selectedBackground || DEFAULT_AVATAR_COLOR;
 
@@ -132,7 +118,7 @@ export function AvatarPicker({
               className="text-sm font-nunito-semibold"
               style={{ color: activeTab === tab ? "#fff" : theme.text }}
             >
-              {tab}
+              {TAB_LABELS[tab]}
             </ThemedText>
           </TouchableOpacity>
         ))}
