+++ frontend/apps/mobile/constants/avatarColors.ts
@@ -1,0 +1,26 @@
+export const AVATAR_COLORS = [
+  "#9CAF7D",
+  "#F5D878",
+  "#F5A85A",
+  "#B5A8E0",
+  "#F0A0BC",
+  "#E07070",
+  "#6B9FD4",
+  "#A8D4E8",
+  "#B0B0B0",
+];
+
+/** Darker shade of each AVATAR_COLOR for use as initials text */
+export const DARK_AVATAR_COLORS: Record<string, string> = {
+  "#9CAF7D": "#5E694B",
+  "#F5D878": "#938248",
+  "#F5A85A": "#936536",
+  "#B5A8E0": "#6D6586",
+  "#F0A0BC": "#906071",
+  "#E07070": "#864343",
+  "#6B9FD4": "#405F7F",
+  "#A8D4E8": "#657F8B",
+  "#B0B0B0": "#6A6A6A",
+};
+
+export const DEFAULT_AVATAR_COLOR = AVATAR_COLORS[3];
