+++ frontend/apps/mobile/i18n/en.json
@@ -157,7 +157,10 @@
     "deleteConfirm": "Are you sure you want to remove this child profile?",
     "requiredFieldsError": "Please fill in all required fields (Name, Birth Date, School ID)",
     "saveError": "Failed to save. Please try again.",
-    "deleteError": "Failed to delete."
+    "deleteError": "Failed to delete.",
+    "changeProfilePicture": "Change Profile Picture",
+    "avatarTabColors": "Colors",
+    "avatarTabAvatar": "Avatar"
   },
   "saved": {
     "title": "Saved",
